% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myscale.R
\name{myscale}
\alias{myscale}
\title{Scale}
\usage{
myscale(input_value, algorithm)
}
\arguments{
\item{input_value}{Numeric vector}

\item{algorithm}{Either "log" for logarithm or "exp" for exponential}
}
\value{
A numeric vector containing the scaled input.
}
\description{
Scale
}
\examples{
scale(10,"log")
}
